requirements.txt
matplotlib>=3.1.0
numpy>=1.18.0
numba>=0.48.0
scipy>=1.4.0
sympy>=1.5.0
py-pde>=0.13.2
